refactor: move http_json_transport into its own file
