+++ apollo-federation/src/sources/connect/mod.rs
@@ -30,9 +30,9 @@ pub use variable::Namespace;
 
 pub use self::models::Connector;
 pub use self::models::EntityResolver;
-pub use self::models::HTTPMethod;
-pub use self::models::HeaderSource;
-pub use self::models::HttpJsonTransport;
+pub use self::models::http_json_transport::HTTPMethod;
+pub use self::models::http_json_transport::HeaderSource;
+pub use self::models::http_json_transport::HttpJsonTransport;
 use crate::schema::position::ObjectFieldDefinitionPosition;
 use crate::schema::position::ObjectOrInterfaceFieldDefinitionPosition;
 use crate::schema::position::ObjectOrInterfaceFieldDirectivePosition;
