+++ apollo-federation/src/sources/connect/models/http_json_transport.rs
@@ -1,0 +1,326 @@
+use std::error::Error;
+use std::fmt::Display;
+use std::fmt::Formatter;
+use std::str::FromStr;
+
+use apollo_compiler::Node;
+use apollo_compiler::ast;
+use apollo_compiler::collections::IndexMap;
+use apollo_compiler::parser::SourceSpan;
+use either::Either;
+use http::HeaderName;
+use url::Url;
+
+use super::super::JSONSelection;
+use super::super::PathSelection;
+use super::super::URLTemplate;
+use super::super::json_selection::ExternalVarPaths;
+use super::super::spec::ConnectHTTPArguments;
+use super::super::spec::SourceHTTPArguments;
+use super::super::spec::versions::AllowedHeaders;
+use super::super::string_template;
+use super::super::variable::Namespace;
+use super::super::variable::VariableReference;
+use crate::error::FederationError;
+use crate::sources::connect::header::HeaderValue;
+use crate::sources::connect::spec::schema::HEADERS_ARGUMENT_NAME;
+use crate::sources::connect::spec::schema::HTTP_HEADER_MAPPING_FROM_ARGUMENT_NAME;
+use crate::sources::connect::spec::schema::HTTP_HEADER_MAPPING_NAME_ARGUMENT_NAME;
+use crate::sources::connect::spec::schema::HTTP_HEADER_MAPPING_VALUE_ARGUMENT_NAME;
+
+#[derive(Clone, Debug)]
+pub struct HttpJsonTransport {
+    pub source_url: Option<Url>,
+    pub connect_template: URLTemplate,
+    pub method: HTTPMethod,
+    pub headers: IndexMap<HeaderName, HeaderSource>,
+    pub body: Option<JSONSelection>,
+}
+
+impl HttpJsonTransport {
+    pub(super) fn from_directive(
+        http: &ConnectHTTPArguments,
+        source: Option<&SourceHTTPArguments>,
+    ) -> Result<Self, FederationError> {
+        let (method, connect_url) = if let Some(url) = &http.get {
+            (HTTPMethod::Get, url)
+        } else if let Some(url) = &http.post {
+            (HTTPMethod::Post, url)
+        } else if let Some(url) = &http.patch {
+            (HTTPMethod::Patch, url)
+        } else if let Some(url) = &http.put {
+            (HTTPMethod::Put, url)
+        } else if let Some(url) = &http.delete {
+            (HTTPMethod::Delete, url)
+        } else {
+            return Err(FederationError::internal("missing http method"));
+        };
+
+        #[allow(clippy::mutable_key_type)]
+        // HeaderName is internally mutable, but we don't mutate it
+        let mut headers = http.headers.clone();
+        for (header_name, header_source) in
+            source.map(|source| &source.headers).into_iter().flatten()
+        {
+            if !headers.contains_key(header_name) {
+                headers.insert(header_name.clone(), header_source.clone());
+            }
+        }
+
+        Ok(Self {
+            source_url: source.and_then(|s| s.base_url.clone()),
+            connect_template: connect_url.parse().map_err(|e: string_template::Error| {
+                FederationError::internal(format!(
+                    "could not parse URL template: {message}",
+                    message = e.message
+                ))
+            })?,
+            method,
+            headers,
+            body: http.body.clone(),
+        })
+    }
+
+    pub(super) fn label(&self) -> String {
+        format!("http: {} {}", self.method.as_str(), self.connect_template)
+    }
+
+    pub(super) fn variables(&self) -> impl Iterator<Item = Namespace> {
+        self.variable_references()
+            .map(|var_ref| var_ref.namespace.namespace)
+    }
+
+    pub(super) fn variable_references(&self) -> impl Iterator<Item = VariableReference<Namespace>> {
+        let url_selections = self.connect_template.expressions().map(|e| &e.expression);
+        let header_selections = self
+            .headers
+            .iter()
+            .flat_map(|(_, source)| source.expressions());
+        url_selections
+            .chain(header_selections)
+            .chain(self.body.iter())
+            .flat_map(|b| {
+                b.external_var_paths()
+                    .into_iter()
+                    .flat_map(PathSelection::variable_reference)
+            })
+    }
+}
+
+/// The HTTP arguments needed for a connect request
+#[derive(Debug, Clone, Copy)]
+pub enum HTTPMethod {
+    Get,
+    Post,
+    Patch,
+    Put,
+    Delete,
+}
+
+impl HTTPMethod {
+    #[inline]
+    pub fn as_str(&self) -> &str {
+        match self {
+            HTTPMethod::Get => "GET",
+            HTTPMethod::Post => "POST",
+            HTTPMethod::Patch => "PATCH",
+            HTTPMethod::Put => "PUT",
+            HTTPMethod::Delete => "DELETE",
+        }
+    }
+}
+
+impl FromStr for HTTPMethod {
+    type Err = String;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s.to_uppercase().as_str() {
+            "GET" => Ok(HTTPMethod::Get),
+            "POST" => Ok(HTTPMethod::Post),
+            "PATCH" => Ok(HTTPMethod::Patch),
+            "PUT" => Ok(HTTPMethod::Put),
+            "DELETE" => Ok(HTTPMethod::Delete),
+            _ => Err(format!("Invalid HTTP method: {s}")),
+        }
+    }
+}
+
+impl Display for HTTPMethod {
+    fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
+        write!(f, "{}", self.as_str())
+    }
+}
+
+#[derive(Clone, Debug)]
+pub enum HeaderSource {
+    From(HeaderName),
+    Value(HeaderValue),
+}
+
+impl HeaderSource {
+    pub(crate) fn expressions(&self) -> impl Iterator<Item = &JSONSelection> {
+        match self {
+            HeaderSource::From(_) => Either::Left(std::iter::empty()),
+            HeaderSource::Value(value) => Either::Right(value.expressions().map(|e| &e.expression)),
+        }
+    }
+}
+
+#[derive(Clone, Debug)]
+pub(crate) struct Header<'a> {
+    pub(crate) name: HeaderName,
+    pub(crate) name_node: &'a Node<ast::Value>,
+    pub(crate) source: HeaderSource,
+    pub(crate) source_node: &'a Node<ast::Value>,
+}
+
+impl<'a> Header<'a> {
+    /// Get a list of headers from the `headers` argument in a `@connect` or `@source` directive.
+    pub(crate) fn from_headers_arg(
+        node: &'a Node<ast::Value>,
+        allowed_headers: &AllowedHeaders,
+    ) -> Vec<Result<Self, HeaderParseError<'a>>> {
+        if let Some(values) = node.as_list() {
+            values
+                .iter()
+                .map(|v| Self::from_single(v, allowed_headers))
+                .collect()
+        } else if node.as_object().is_some() {
+            vec![Self::from_single(node, allowed_headers)]
+        } else {
+            vec![Err(HeaderParseError::Other {
+                message: format!("`{HEADERS_ARGUMENT_NAME}` must be an object or list of objects"),
+                node,
+            })]
+        }
+    }
+
+    /// Build a single [`Self`] from a single entry in the `headers` arg.
+    fn from_single(
+        node: &'a Node<ast::Value>,
+        allowed_headers: &AllowedHeaders,
+    ) -> Result<Self, HeaderParseError<'a>> {
+        let mappings = node.as_object().ok_or_else(|| HeaderParseError::Other {
+            message: "the HTTP header mapping is not an object".to_string(),
+            node,
+        })?;
+        let name_node = mappings
+            .iter()
+            .find_map(|(name, value)| {
+                (*name == HTTP_HEADER_MAPPING_NAME_ARGUMENT_NAME).then_some(value)
+            })
+            .ok_or_else(|| HeaderParseError::Other {
+                message: format!("missing `{HTTP_HEADER_MAPPING_NAME_ARGUMENT_NAME}` field"),
+                node,
+            })?;
+        let name = name_node
+            .as_str()
+            .ok_or_else(|| format!("`{HTTP_HEADER_MAPPING_NAME_ARGUMENT_NAME}` is not a string"))
+            .and_then(|name_str| {
+                HeaderName::try_from(name_str)
+                    .map_err(|_| format!("the value `{name_str}` is an invalid HTTP header name"))
+            })
+            .map_err(|message| HeaderParseError::Other {
+                message,
+                node: name_node,
+            })?;
+
+        if allowed_headers.header_name_is_reserved(&name) {
+            return Err(HeaderParseError::Other {
+                message: format!("header '{name}' is reserved and cannot be set by a connector"),
+                node: name_node,
+            });
+        }
+
+        let from = mappings
+            .iter()
+            .find(|(name, _value)| *name == HTTP_HEADER_MAPPING_FROM_ARGUMENT_NAME);
+        let value = mappings
+            .iter()
+            .find(|(name, _value)| *name == HTTP_HEADER_MAPPING_VALUE_ARGUMENT_NAME);
+
+        match (from, value) {
+            (Some(_), None) if allowed_headers.header_name_allowed_static(&name) => {
+                Err(HeaderParseError::Other{ message: format!(
+                    "header '{name}' can't be set with `{HTTP_HEADER_MAPPING_FROM_ARGUMENT_NAME}`, only with `{HTTP_HEADER_MAPPING_VALUE_ARGUMENT_NAME}`"
+                ), node: name_node})
+            }
+            (Some((_, from_node)), None) => {
+                from_node.as_str()
+                    .ok_or_else(|| format!("`{HTTP_HEADER_MAPPING_FROM_ARGUMENT_NAME}` is not a string"))
+                    .and_then(|from_str| {
+                        HeaderName::try_from(from_str).map_err(|_| {
+                            format!("the value `{from_str}` is an invalid HTTP header name")
+                        })
+                    })
+                    .map(|from| Self {
+                        name,
+                        name_node,
+                        source: HeaderSource::From(from),
+                        source_node: from_node,
+                    })
+                    .map_err(|message| HeaderParseError::Other{ message, node: from_node})
+            }
+            (None, Some((_, value_node))) => {
+                value_node
+                    .as_str()
+                    .ok_or_else(|| HeaderParseError::Other{ message: format!("`{HTTP_HEADER_MAPPING_VALUE_ARGUMENT_NAME}` field in HTTP header mapping must be a string"), node: value_node})
+                    .and_then(|value_str| {
+                        value_str
+                            .parse::<HeaderValue>()
+                            .map_err(|err| HeaderParseError::ValueError {err, node: value_node})
+                    })
+                    .map(|value| Self {
+                        name,
+                        name_node,
+                        source: HeaderSource::Value(value),
+                        source_node: value_node,
+                    })
+            }
+            (None, None) => {
+                Err(HeaderParseError::Other {
+                    message: format!("either `{HTTP_HEADER_MAPPING_FROM_ARGUMENT_NAME}` or `{HTTP_HEADER_MAPPING_VALUE_ARGUMENT_NAME}` must be set"),
+                    node,
+                })
+            },
+            (Some((from_name, _)), Some((value_name, _))) => {
+                Err(HeaderParseError::ConflictingArguments {
+                    message: format!("`{HTTP_HEADER_MAPPING_FROM_ARGUMENT_NAME}` and `{HTTP_HEADER_MAPPING_VALUE_ARGUMENT_NAME}` can't be set at the same time"),
+                    from_location: from_name.location(),
+                    value_location: value_name.location(),
+                })
+            }
+        }
+    }
+}
+
+#[derive(Debug)]
+pub(crate) enum HeaderParseError<'a> {
+    ValueError {
+        err: string_template::Error,
+        node: &'a Node<ast::Value>,
+    },
+    /// Both `value` and `from` are set
+    ConflictingArguments {
+        message: String,
+        from_location: Option<SourceSpan>,
+        value_location: Option<SourceSpan>,
+    },
+    Other {
+        message: String,
+        node: &'a Node<ast::Value>,
+    },
+}
+
+impl Display for HeaderParseError<'_> {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::ConflictingArguments { message, .. } | Self::Other { message, .. } => {
+                write!(f, "{}", message)
+            }
+            Self::ValueError { err, .. } => write!(f, "{err}"),
+        }
+    }
+}
+
+impl Error for HeaderParseError<'_> {}
